chore: move MainMenu content to App, delete MainMenu
